incorporate review comments
