+++ docs/docs/06-cli.md
@@ -18,4 +18,4 @@ replacing `<path-to-file>` with the path to a file containing MemoryViz-compatib
 
 ## Output
 
-The output is an SVG image generated by MemoryViz and the image is saved in the same folder in which the CLI is called. The name of the SVG will be the same as that of the inputted file (i.e., if the inputted file is `david-is-cool.json`, the output will be `david-is-cool.svg`).
+The output is an SVG image generated by MemoryViz and the image is saved in the current working directory. The name of the SVG will be the same as that of the inputted file (i.e., if the inputted file is `david-is-cool.json`, the output will be `david-is-cool.svg`).
